move plotutil to ctrlplot
